show task duration on /tasks and /tasks/taskSlug pages
