+++ src/components/BulkDeployments/index.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 
 import CancelDeployment from 'components/CancelDeployment';
-import { getDeploymentDuration } from 'components/Deployment';
 import { formatString } from 'components/DeploymentsByFilter';
 import HoverTag from 'components/HoverTag';
 import DeploymentLink from 'components/link/Deployment';
 import DeploymentsLink from 'components/link/Deployments';
 import ProjectLink from 'components/link/Project';
+import { getProcessDuration } from 'lib/util';
 import moment from 'moment';
 
 import { BulkDeploymentsDataTable, BulkDeploymentsHeader } from './StyledBulkDeployments';
@@ -79,7 +79,7 @@ const BulkDeployments: FC<BulkDeploymentsProps> = ({ deployments }) => (
               <HoverTag text={deployment.buildStep} maxWidth="160px" />
             )}
           </div>
-          <div className="duration">{getDeploymentDuration(deployment)}</div>
+          <div className="duration">{getProcessDuration(deployment)}</div>
           <div>
             {['new', 'pending', 'queued', 'running'].includes(deployment.status) && (
               <CancelDeployment deployment={deployment} afterText="Cancelled" beforeText="Cancel" />
