+++ src/lib/query/EnvironmentWithTask.js
@@ -18,6 +18,8 @@ export default gql`
         taskName
         status
         created
+        started
+        completed
         service
         logs
         adminOnlyView
