+++ src/lib/util.js
@@ -1,3 +1,5 @@
+import moment from 'moment';
+import momentDurationFormatSetup from 'moment-duration-format';
 import * as R from 'ramda';
 
 export const queryStringToObject = R.pipe(
@@ -19,4 +21,13 @@ export const debounce = (fn, delay) => {
       fn.call(null, val);
     }, delay);
   };
+};
+
+export const getProcessDuration = process => {
+  const processStart = process.started || process.created;
+  const durationStart = (processStart && moment.utc(processStart)) || moment.utc();
+  const durationEnd = (process.completed && moment.utc(process.completed)) || moment.utc();
+  const duration = moment.duration(durationEnd - durationStart).format('HH[hr] mm[m] ss[sec]');
+
+  return duration;
 };
