+++ src/webpack.shared-config.js
@@ -4,6 +4,7 @@ module.exports = {
   extensions: ['.js', '.tsx'],
   alias: {
     components: path.join(__dirname, 'components'),
+    styles: path.join(__dirname, 'styles'),
     layouts: path.join(__dirname, 'layouts'),
     lib: path.join(__dirname, 'lib'),
     pages: path.join(__dirname, 'pages'),
